Split up new DLC code into six object files
